move the code from BinaryLogImpl to io.grpc.BinaryLog
